Open application upon build success
If the build succeeds then the application
exe will run. This allows for very quick
testing of changes.
The build script will now print a
message to the screen if the variables
MSVSDIR or VCRUN are incorrectly set.
There are also some small fixes to the
README.md correcting some hastily written
instructions.